# internal/api/docs/swagger.yaml
info:
  contact:
    url: https://github.com/httplock/httplock
  description: HTTP reproducible proxy server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: httplock API
  version: "0.1"
paths:
  /api/ca:
    get:
      description: returns the public CA in PEM format
      produces:
      - application/text
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Get CA
  /api/root/:
    get:
      description: Lists the roots
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "500":
          description: Internal Server Error
      summary: Root List
  /api/root/{root}/dir:
    get:
      description: Lists a directory in a root
      parameters:
      - description: root hash or uuid
        in: path
        name: root
        required: true
        type: string
      - description: path to list
        in: query
        items:
          type: string
        name: path
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Root Dir
  /api/root/{root}/export:
    get:
      description: Exports a hash, returning a tar+gz
      parameters:
      - description: hash
        in: path
        name: root
        required: true
        type: string
      produces:
      - application/x-gtar
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Root export
  /api/root/{root}/file:
    get:
      description: Get file contents in a root
      parameters:
      - description: root hash or uuid
        in: path
        name: root
        required: true
        type: string
      - description: path of file
        in: query
        items:
          type: string
        name: path
        type: array
      produces:
      - application/octet-stream
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Root File
  /api/root/{root}/import:
    put:
      consumes:
      - application/x-gtar
      description: Imports a root hash from a tar+gz
      parameters:
      - description: hash
        in: path
        name: root
        required: true
        type: string
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Root import
  /api/token:
    post:
      description: returns a new uuid for recording a session
      parameters:
      - description: hash used to initialize the response cache
        in: query
        name: hash
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "500":
          description: Internal Server Error
      summary: Token create
  /api/token/{id}:
    delete:
      description: returns a new uuid for recording a session
      parameters:
      - description: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Token delete
  /api/token/{id}/save:
    post:
      description: Saves a uuid token, returning an immutable hash
      parameters:
      - description: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Token save
swagger: "2.0"